refactor (Activity): 액티비티 별로 패키지 구분
